+++ src/main/java/net/geforcemods/securitycraft/blockentities/KeypadBlockEntity.java
@@ -2,7 +2,6 @@ package net.geforcemods.securitycraft.blockentities;
 
 import java.util.UUID;
 
-import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.ILockable;
 import net.geforcemods.securitycraft.api.IPasscodeProtected;
 import net.geforcemods.securitycraft.api.Option;
@@ -15,6 +14,7 @@ import net.geforcemods.securitycraft.api.Option.SignalLengthOption;
 import net.geforcemods.securitycraft.api.Option.SmartModuleCooldownOption;
 import net.geforcemods.securitycraft.blocks.KeypadBlock;
 import net.geforcemods.securitycraft.misc.ModuleType;
+import net.geforcemods.securitycraft.util.BlockUtils;
 import net.geforcemods.securitycraft.util.PasscodeUtils;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
@@ -23,17 +23,6 @@ import net.minecraft.nbt.NBTTagCompound;
 public class KeypadBlockEntity extends DisguisableBlockEntity implements IPasscodeProtected, ILockable {
 	private byte[] passcode;
 	private UUID saltKey;
-	private BooleanOption isAlwaysActive = new BooleanOption("isAlwaysActive", false) {
-		@Override
-		public void toggle() {
-			super.toggle();
-
-			if (!isDisabled()) {
-				world.setBlockState(pos, world.getBlockState(pos).withProperty(KeypadBlock.POWERED, get()));
-				world.notifyNeighborsOfStateChange(pos, SCContent.keypad, false);
-			}
-		}
-	};
 	private BooleanOption sendAllowlistMessage = new SendAllowlistMessageOption(false);
 	private BooleanOption sendDenylistMessage = new SendDenylistMessageOption(true);
 	private IntOption signalLength = new SignalLengthOption(this::getPos, 60);
@@ -83,14 +72,11 @@ public class KeypadBlockEntity extends DisguisableBlockEntity implements IPassco
 
 	@Override
 	public void onOptionChanged(Option<?> option) {
-		if (option.getName().equals("disabled")) {
-			boolean isDisabled = ((BooleanOption) option).get();
+		if ((option.getName().equals(disabled.getName()) && ((BooleanOption) option).get() || option.getName().equals(signalLength.getName()))) {
 			IBlockState state = world.getBlockState(pos);
 
-			if (isDisabled && state.getValue(KeypadBlock.POWERED))
-				world.setBlockState(pos, state.withProperty(KeypadBlock.POWERED, false));
-			else if (!isDisabled && isAlwaysActive.get())
-				world.setBlockState(pos, state.withProperty(KeypadBlock.POWERED, true));
+			world.setBlockState(pos, state.withProperty(KeypadBlock.POWERED, false));
+			BlockUtils.updateIndirectNeighbors(world, pos, state.getBlock());
 		}
 	}
 
@@ -145,7 +131,7 @@ public class KeypadBlockEntity extends DisguisableBlockEntity implements IPassco
 	@Override
 	public Option<?>[] customOptions() {
 		return new Option[] {
-				isAlwaysActive, sendAllowlistMessage, sendDenylistMessage, signalLength, disabled, smartModuleCooldown
+				sendAllowlistMessage, sendDenylistMessage, signalLength, disabled, smartModuleCooldown
 		};
 	}