+++ src/main/java/net/geforcemods/securitycraft/blockentities/RiftStabilizerBlockEntity.java
@@ -42,7 +42,7 @@ import net.minecraftforge.common.util.Constants.NBT;
 import net.minecraftforge.fml.common.registry.ForgeRegistries;
 
 public class RiftStabilizerBlockEntity extends DisguisableBlockEntity implements ITickable, ILockable, IToggleableEntries<TeleportationType> {
-	private final IntOption signalLength = new IntOption(this::getPos, "signalLength", 60, 5, 400, 5, true); //20 seconds max
+	private final IntOption signalLength = new IntOption(this::getPos, "signalLength", 60, 0, 400, 5, true); //20 seconds max
 	private final IntOption range = new IntOption(this::getPos, "range", 5, 1, 15, 1, true);
 	private final DisabledOption disabled = new DisabledOption(false);
 	private IgnoreOwnerOption ignoreOwner = new IgnoreOwnerOption(true);