+++ src/main/java/net/geforcemods/securitycraft/blocks/RiftStabilizerBlock.java
@@ -52,8 +52,10 @@ public class RiftStabilizerBlock extends DisguisableBlock {
 
 	@Override
 	public void updateTick(World world, BlockPos pos, IBlockState state, Random random) {
-		world.setBlockState(pos, state.withProperty(POWERED, false));
-		BlockUtils.updateIndirectNeighbors(world, pos, this);
+		if (state.getValue(POWERED)) {
+			world.setBlockState(pos, state.withProperty(POWERED, false));
+			BlockUtils.updateIndirectNeighbors(world, pos, this);
+		}
 	}
 
 	@Override